restructure repository folder
