+++ internal/repository/kelas.go
@@ -1,0 +1,1 @@
+package repository
