+++ internal/service/kelas.go
@@ -1,0 +1,1 @@
+package service
