+++ internal/service/user_service/user_activate_deactivate.go
@@ -1,0 +1,1 @@
+package user_service
